# swagger/v2/swagger.yaml
swagger: '2.0'
info:
  title: METAX API
  version: v2
host: __METAX_ENV_DOMAIN__
schemes:
  - https
basePath: ''
produces:
  - application/json
paths:
...
  /rest/v2/files/{PID}/XML:
    get:
      summary: Get XML metadata from file
      produces:
        - application/xml
        - application/json
      parameters:
        - name: PID
          in: path
          description: Persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: Namespace of the schema. If omitted, returns a list of xml metadata namespaces associated with the file.
          type: string
      responses:
        '200':
          description: Single xml metadata (application/xml), or a list of namespaces (application/json)
        '404':
          description: File not found. Also when not authenticated request
      tags:
        - File API
    post:
      summary: Create new XML metadata
      consumes:
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - name: PID
          in: path
          description: Persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: Namespace of the schema. If a xml metadata already existed with given namespace, an error is returned
          required: true
          type: string
        - name: body
          in: body
          description: The XML to create.
          required: true
        - $ref: "#/parameters/dryrun"
      responses:
        '201':
          description: Return created XML
        '400':
          description: Parameters contained errors, response includes details (application/json)
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - File API
    put:
      summary: Update XML metadata
      consumes:
        - application/xml
      parameters:
        - name: PID
          in: path
          description: Persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: Namespace of the schema
          required: true
          type: string
        - name: body
          in: body
          description: The XML to update
          required: true
        - $ref: "#/parameters/dryrun"
      responses:
        '204':
          description: XML saved
        '400':
          description: Parameters contained errors, response includes details
        '403':
          description: Forbidden. Must have permission for resource
        '404':
          description: File or namespace does not exist
      tags:
        - File API
    delete:
      summary: Delete XML metadata
      parameters:
        - name: PID
          in: path
          description: Persistent ID of file OR the internal pk
          required: true
          type: string
        - name: namespace
          in: query
          description: Namespace of the schema
          required: true
          type: string
        - $ref: "#/parameters/dryrun"
      responses:
        '204':
          description: XML is deleted
        '400':
          description: Parameters contained errors, response includes details
        '403':
          description: Forbidden. Must have permission for resource
        '404':
          description: File or namespace does not exist
      tags:
        - File API
  /rest/v2/files/datasets:
    post:
      summary: Get datasets where files belong to
      description: |
        Note&#58; The method is invoked using POST, because there are limits to length of query
        parameters in GET. Also, some clients forcibly shove parameters in body in GET
        requests to query parameters, so using POST instead is more guaranteed to work.
      produces:
        - application/json
      parameters:
        - name: body
          in: body
          description: A list of file id's (integers) or identifiers (strings).
          required: true
          schema:
            $ref: '#/definitions/StringList'
        - $ref: "#/parameters/dryrun"
        - name: keys=datasets
          in: query
          description: A list of dataset id's (integers) or identifiers (strings).
                      Returns the files that belong to datasets. If dataset doesn't contain files,
                      an empty list is returned.
          required: false
          schema:
            $ref: '#/definitions/StringList'
        - name: keys=files
          in: query
          description: A list of file id's (integers) or identifiers (strings).
                        Returns the datasets that the file belongs to. If a file doesn't belong to any dataset,
                        an empty list is returned.
          required: false
          schema:
            $ref: '#/definitions/StringList'
        - name: keysonly
          in: query
          description: Returns those files from original input that belong to a dataset,
                        or in case of keys=datasets, returns those datasets from original list of datasets
                        that have files.
          required: false
          type: boolean
      responses:
        '200':
          description: A list of preferred_identifiers. if the files were not found to belong to any datasets, an empty list is returned
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - File API
...
  /rest/v2/datasets:
    # some of the parameters and returned fields are for TPAS usage only
    get:
      summary: List datasets
      parameters:
        - name: latest
          in: query
          description: Only return latest versions
          required: false
          type: boolean
        - name: owner_id
          in: query
          description: Id of the person who owns the record in metax
          required: false
          type: string
        - name: user_created
          in: query
          description: Id of the person who created the record in metax
          required: false
          type: string
        - name: curator
          in: query
          description: Curator identifier (field research_dataset-> curator-> identifier)
          required: false
          type: string
        - name: preferred_identifier
          in: query
          description: |
            Use given preferred_identifier as filter. Returns a best guess of the record (always a single record!), by
            looking from Fairdata catalogs first, then from harvested catalogs. If the value of preferred_identifier
            contains special characters such as ampersands, the value has to be urlencoded.
          required: false
          type: string
        - name: state
          in: query
          description: Filter by state field. Possible values are&#58; draft, publihed.
          required: false
          type: string
        - name: preservation_state
          in: query
          description: TPAS state (field preservation_state). Multiple states using OR-logic are queriable in the same request, e.g. state=5,6. See valid values from http://iow.csc.fi/model/mrd/CatalogRecord/ field preservation_state
          required: false
          type: string
        - name: metadata_owner_org
          in: query
          description: Filter by dataset field metadata_owner_org
          required: false
          type: string
        - name: metadata_provider_user
          in: query
          description: Filter by dataset field metadata_provider_user
          required: false
          type: string
        - name: contract_org_identifier
          in: query
          description: Filter by dataset contract.contract_json.organization.organization_identifier. Restricted to permitted users.
          required: false
          type: string
        - name: pas_filter
          in: query
          description: A specific filter that targets the following fields; research_dataset['title'], research_dataset['curator'][n]['name'], contract['contract_json']['title']. Restricted to permitted users.
          required: false
          type: string
        - name: editor
          in: query
          description: Identifier of the editor used to modify the record, i.e. qvain.
          type: string
        - name: data_catalog
          in: query
          description: Filter by data catalog urn identifier
          type: string
        - name: offset
          in: query
          description: Offset for paging
          required: false
          type: integer
        - name: limit
          in: query
          description: Limit for paging
          required: false
          type: integer
          default: 10
        - name: ordering
          in: query
          description: Specify ordering of results by fields. Accepts a list of field names separated by a comma. Ordering can be reversed by prefixing field name with a '-' char.
          required: false
          type: string
        - name: actor_filter
          in: query
          description: |
            Actor_filters are a collection of filter parameters for filtering according to the name
            of creator, curator, publisher or rights_holder actors. Actor type must be defined as
            a suffix in the filter name ('_person' or '_organization'). Actor type '_organization' finds matches
            from "is_member_of" -field if actor is a person. Multiple actor_filters can be applied simultaneously (AND)
            or separately (OR) by using 'condition_separator'. Default separator is AND.

            Examples&#58;

            '?creator_person=person_name'
            '?publisher_organization=some organisation&creator_person=person_name&condition_separator=or'
          required: false
          type: string
        - name: api_version
          in: query
          description: Returns datasets that can be edited with certain api version. Possible values are 1 and 2
          type: integer
          required: false
        - $ref: "#/parameters/include_user_metadata"
        - $ref: "#/parameters/fields"
      responses:
        "200":
          description: Successful operation, returns a list of datasets containing full dataset objects including their data catalog and contract information. When using query parameters, search result can be an empty list.
          schema:
            $ref: '#/definitions/CatalogRecord'
      tags:
        - Dataset API
    post:
      summary: Create new dataset metadata
      description: |
        # research_dataset identifiers
        The fields <b>metadata_version_identifier</b> and <b>preferred_identifier</b> are generally always generated server-side. <b>Exception</b>: In harvested catalogs, preferred_identifier value can be provided by the user.
        # research_dataset read-only fields
        - metadata_version_identifier
        - preferred_identifier (except harvested catalogs)
        - total_files_byte_size (generated from files in the db according to files and directories listed in the dataset metadata)
        - total_remote_resources_byte_size (generated from entries in remote_resources)
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "Object or a list of objects to create."
          required: true
          schema:
            $ref: '#/definitions/CatalogRecord'
        - in: query
          name: draft
          description: Create new dataset into draft state. The dataset has to be separately published using the designated RPC API.
          type: boolean
          required: false
          default: false
        - $ref: "#/parameters/allowed_projects"
        - $ref: "#/parameters/dryrun"
        - $ref: "#/parameters/include_user_metadata"
      responses:
        '201':
          description: Returns the created object, or if a list was given, a list of objects and errors.
          schema:
            $ref: '#/definitions/CatalogRecord'
        '400':
          description: Parameters contained errors, response includes details.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - Dataset API
    put:
      summary: Bulk update
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "A list of objects to update."
          required: true
          schema:
            $ref: '#/definitions/CatalogRecordList'
        - $ref: "#/parameters/allowed_projects"
        - $ref: "#/parameters/dryrun"
      responses:
        '200':
          description: Sucessful operation. Return values include a list of errors, if any.
        '400':
          description: All updates failed. A list of errors is returned.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - Dataset API
    patch:
      summary: Bulk update partial
      description: |
        The payload must include a field that can be used to identify the resource being updated. Acceptable identifier fields are: <b>id, identifier</b>
      consumes:
        - application/json
      parameters:
        - in: "body"
          name: "body"
          description: "A list of (partial) objects to update."
          required: true
          schema:
            $ref: '#/definitions/CatalogRecordList'
        - $ref: "#/parameters/allowed_projects"
        - $ref: "#/parameters/dryrun"
      responses:
        '200':
          description: Some or all objects were updated. Return values contain list of full updated objects, and may include a list of errors.
        '400':
          description: All updates failed. A list of errors is returned.
        '403':
          description: Forbidden. Must have permission for resource
      tags:
        - Dataset API
...
  /rpc/files/flush_project:
    post:
      summary: Permanently DELETES an entire project's files and directories.
      description: Permanently DELETES an entire project's files and directories. Does not check file association with datasets! Not meant for active production use!!.
      parameters:
        - name: identifier
          in: query
          description: Identifier of the catalog record to be fixed.
          required: true
          type: string
      responses:
        '200':
          description: Successful operation, returns information about the new dataset version.
        '400':
          description: API currently allowed only in test environments.
      tags:
        - File RPC
...
parameters:
  dryrun:
    name: dryrun
    in: query
    description: Execute the operation normally, returning the same response as normally, but do not save anything to db, or publish anything to remote services.
    required: false
    type: boolean
  allowed_projects:
    name: allowed_projects
    in: query
    description: Ensures that content is part of intended projects.
    required: false
    type: string
...
  include_user_metadata:
    name: include_user_metadata
    in: query
    description: Include user metadata for files (dataset-specific metadata) (fields research_dataset.files and research_dataset.directories)
    required: false
    type: boolean
    default: false
...
  fields:
    name: fields
    in: query
    description: Comma separated list of fields that is returned. Note that nested fields are not supported.
    required: false
    type: string
definitions:
...
  CatalogRecord:
    $ref: 'https://__METAX_ENV_DOMAIN__/apischemas/v2/catalogrecord.json#/definitions/CatalogRecord'
...
  StringList:
    type: array
    items:
      type: string
      example: "value"
  CatalogRecordList:
    type: array
    items:
      $ref: '#/definitions/CatalogRecord'
...